Remove user resource management endpoints
Users should not be manageable through the API. Or at least not in the
basic way is is now. This must be done with an intelligent system to
prevent misuse of user permissions. (Needs e.g. flows for deleting user
with strikes/warnings instead of simply dropping them from database).